Added Amphenol 12401548E4-2A
The hybrid version of Amphenol 12401610E4-2A.

Datasheet (link on this page, can't seem to get a direct link):
https://www.amphenolcanada.com/StockAvailabilityPrice.aspx?From=&PartNum=12401548E4%7e2A

The annular rings of the B pins in the footprint are only 0.125 mm,
which is smaller than KLC allows.  I believe this matches Amphenol's
drawing (though it doesn't appear they explicitly dimensioned the
annular rings, the PCB layout drawing is to scale, and the annular
rings certainly look like 0.125 mm).